requirements.txt
PyYAML==6.0
ccxt==1.95.25
